Refactored BasecampConfig object and added new test for correctness
New test to ensure all ways of calling the main Basecamp3 object work.
BasecampConfig is now abstract.
BasecampMemoryConfig and BasecampFileConfig handle the problem with save() if a file was not used.
pytz added to ensure correct timezone conversion when determining if a token is expired
Instead of using the default logger, a package-level one has been added.